requirements.txt
requests~=2.24.0
aiogram~=2.10.1
loguru~=0.5.3
dialogflow~=1.1.0